First version that seems to work as expected.
Ran many tests with another sample board which cannot be resolved.
It finds a little over 4,000 boards before giving up. I made sure
that all the boards tested were actually unique and they all are.
None of them are all zeroes (actually, not that many tiles get
removed in the first place).

In any event, I think this is it.

This version also does not generate all the boards as an image.
You can run the tool to generate the image of a specific board
if you'd like.